fix(ui): show @mention autocomplete when editing a message
## Problem

The @mention autocomplete popup worked in the new-message composer but
not when editing an existing message. The mention machinery (query
detection, candidate filtering, caret math, dropdown rendering) lived
entirely inside `MessageInput` and was hardcoded to the composer's
`#message-input` textarea, so the inline edit form in
`MessageGroupComponent` had none of it — typing `@` while editing did
nothing.

## Approach

Extract the mention machinery into a shared
`components/conversation/mention.rs` module, parameterised by the target
textarea's DOM `id` so the same code drives both call sites instead of
being duplicated (which would drift):

- `detect_mention_query`, `filter_mention_candidates`,
  `duplicate_candidate_names`, UTF-16↔byte caret helpers — moved verbatim
  (logic unchanged), with their unit tests.
- `update_mention_from_input`, `handle_mention_keydown`,
  `apply_mention_selection` — small wrappers that take a textarea `id`.
- `MentionDropdown` component — the dropdown markup, single-sourced.

The composer is rewired to call the shared module; its rendered DOM is
unchanged (verified: composer popup still works). The edit textarea in
`MessageGroupComponent` is wrapped in a `relative` container, given a
unique `edit-msg-<id>` id, and wired to the same oninput/onkeydown/
dropdown path. Its mentionable-member list is derived from the
`member_names` map the component already receives (self excluded, sorted
by name), so no new prop plumbing. Escape closes the dropdown first
(then a second Escape cancels the edit); Enter/Tab accept a candidate
without saving; clicking elsewhere dismisses via onfocusout.

UI-only change — no delegate/contract/common WASM touched, so no
migration entry is required.

## Testing

- Relocated mention unit tests pass (`cargo test -p river-ui --bins`,
  348 passed).
- Headless Playwright run against the example-data build confirms:
  composer popup still shows (regression); edit popup now shows and is
  on-screen/unclipped; mouse-pick and keyboard (ArrowDown+Enter) both
  insert the correct mention token and close the dropdown; Enter-accept
  does not save the edit; Escape closes the dropdown without cancelling.

[AI-assisted - Claude]
